+++ src/Streams/Readable.php
@@ -3,10 +3,13 @@ declare(strict_types = 1);
 
 namespace Innmind\Stream\Streams;
 
-use Innmind\Stream\Readable as Read;
+use Innmind\Stream\{
+    Capabilities,
+    Readable as Read,
+};
 use Innmind\Url\Path;
 
-final class Readable
+final class Readable implements Capabilities\Readable
 {
     private function __construct()
     {
@@ -23,5 +26,10 @@ final class Readable
     public function open(Path $path): Read
     {
         return Read\Stream::open($path);
+    }
+
+    public function acquire($resource): Read
+    {
+        return Read\Stream::of($resource);
     }
 }