+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- `Innmind\Stream\Capabilities` (implemented by already existing `Innmind\Stream\Streams`)
+- `Innmind\Stream\Streams::fromAmbientAuthority()`
+- `Innmind\Stream\Stream::resource(): resource`
+
+### Deprecated
+
+- `Innmind\Stream\Streams::of()` use `::fromAmbientAuthority()` instead
+
+### Removed
+
+- `Innmind\Stream\Selectable`
+
 ## 3.3.0 - 2022-12-04
 
 ### Added